changed back bot name to Duke
